feat(dash): add delay box and adjust help
